+++ src/app/posts/[id]/page.tsx
@@ -4,6 +4,7 @@ import { notFound } from "next/navigation"
 import { TagList } from "@/components/tag/list"
 import "katex/dist/katex.min.css"
 import "highlight.js/styles/base16/google-dark.min.css"
+import { formatDate } from "@/lib/date"
 
 type Props = { params: Promise<{ id: string }> }
 
@@ -25,26 +26,18 @@ export async function generateMetadata(params: Props) {
     }
 }
 
-const options: Intl.DateTimeFormatOptions = {
-    weekday: "long",
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-    timeZone: "Asia/Tokyo",
-}
-
 export default async function Post(props: Props) {
     const post = await getPost(props)
-    const createdAt = new Date(post.createdAt)
-    const updatedAt = new Date(post.updatedAt)
+    const createdAt = formatDate(post.createdAt)
+    const updatedAt = formatDate(post.updatedAt)
     const content = await markReact(post.content)
     return (
         <div>
             <h1 className="mb-4 text-4xl font-bold">{post.title}</h1>
             <TagList className="mb-4 text-lg" tags={post.tags} isLink />
             <div className="mb-8 md:flex md:gap-8">
-                <p>作成日 {createdAt.toLocaleDateString("ja-JP", options)}</p>
-                <p>更新日 {updatedAt.toLocaleDateString("ja-JP", options)}</p>
+                <p>作成日 {createdAt}</p>
+                <p>更新日 {updatedAt}</p>
             </div>
             <div className="prose prose-lg prose-slate max-w-none dark:prose-invert">
                 {content}
