+++ src/components/post/card.tsx
@@ -1,18 +1,29 @@
 import Link from "next/link.js"
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import {
+    Card,
+    CardContent,
+    CardFooter,
+    CardHeader,
+    CardTitle,
+} from "@/components/ui/card"
 import { Skeleton } from "@/components/ui/skeleton"
 import { TagList } from "@/components/tag/list"
 import { Post, Tag } from "@prisma/client"
+import { formatDate } from "@/lib/date"
 
 export function PostCard({ post }: { post: Post & { tags: Tag[] } }) {
     return (
         <Link href={`/posts/${post.id}`}>
-            <Card className="h-36 w-64 hover:border-foreground">
+            <Card className="flex h-36 w-64 flex-col justify-between hover:border-foreground">
                 <CardHeader>
                     <CardTitle className="truncate">{post.title}</CardTitle>
-                    <TagList tags={post.tags} />
                 </CardHeader>
-                <CardContent className="truncate">{post.content}</CardContent>
+                <CardContent className="pb-2">
+                    <TagList tags={post.tags} />
+                </CardContent>
+                <CardFooter>
+                    <p>{formatDate(post.createdAt)}</p>
+                </CardFooter>
             </Card>
         </Link>
     )
